Refactor to simplify the installer client and remove unused methods
This moves the token and secret handling into the client as well as the
refresh logic around the API token.

The client is only used for a single call so it's safe to just get a
token each time. Additionally each token we get will likely be for a
different user in the future so we can't cache the client in the server
either way.