MGMT-20908: Forward authorization
This patch changes the MCP server so that if the `Authorization` header
contains a bearer token it will forward it to the assisted service REST
API.

To existing behaviour to get the use the offline token from the
environment variable `OFFLINE_TOKEN` is preserved when there is no
`Authorization` header in the request.

Note that in order to test this it is necessary to have a MCP client
that is able to send the `Authorization` header. I did so making a small
change to the `mcp` command line tool (from the
https://github.com/f/mcptools project) so that it reads the token from a
`TOKEN` environment variable:

```diff
diff --git a/cmd/mcptools/commands/utils.go b/cmd/mcptools/commands/utils.go
index f3a9be0..19918f2 100644
--- a/cmd/mcptools/commands/utils.go
+++ b/cmd/mcptools/commands/utils.go
@@ -5,6 +5,7 @@ import (
        "context"
        "encoding/json"
        "fmt"
+       "os"
        "strings"
        "time"

@@ -47,7 +48,13 @@ var CreateClientFunc = func(args []string, _ ...client.ClientOption) (*client.Cl
        var err error

        if len(args) == 1 && IsHTTP(args[0]) {
-               c, err = client.NewSSEMCPClient(args[0])
+               token := os.Getenv("TOKEN")
+               c, err = client.NewSSEMCPClient(
+                       args[0],
+                       client.WithHeaders(map[string]string{
+                               "Authorization": fmt.Sprintf("Bearer %s", token),
+                       }),
+               )
                if err != nil {
                        return nil, err
                }
```

Then I can put the access token in that environment variable, and call a
tool. For example, start the server without the offline token:

```shell
$ unset OFFLINE_TOKEN
$ uv run server.py
```

Then get the token and call some tool:

```shell
$ ocm login --use-auth-code
$ TOKEN=$(ocm token) mcp call list_clusters http://localhost:8000/sse | jq
[
  {
    "name": "sno",
    "id": "aef5bb4e-397d-4e21-a776-5f1405b2cf10",
    "openshift_version": "4.19.0",
    "status": "pending-for-input"
  }
]
```

Related: https://issues.redhat.com/browse/MGMT-20908
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>