Create a dedicated tool for getting the ISO download URL
The information in the infraenv is generally not of much use to the
user. This is why the entire concept is hidden from them in the SaaS UI.
In order to be able to support getting the download URL for clusters
created outside of the current session and to further hide the infraenv
concept this commit changes the infraenv_info tool into
cluster_iso_download_url which takes just a cluster ID as input.

While a cluster can have multiple infraenvs, this is currently not an
expected flow in the SaaS so we simply return all that we find without
additional context. In the future if this MCP server is required to
support multi-arch deployments (the most common use-case for multiple
infraenvs) then maybe more information will need to accompany the
multiple URLs.

https://issues.redhat.com/browse/MGMT-21040