Refresh the iso download url when querying for it
The download URL in the infraenv will be valid for a few hours after the
cluster is created, but once the URL expires the dedicated URL API is
required to refresh it. This commit updates the cluster_iso_download_url
tool to just use that API all the time instead of pulling the url and
expiration out of the infraenv.